Updated Dockerfile, Docker-compose and main.py
Should be ready to test if it works in docker environment. Updated the ports to match between app initialisation in main.py and service registration in docker-compose. Dockerfile is the same conda activation though whether it works is unknown. Also added host specification of 0.0.0.0 to webserver run command as precaution.