Updated Dockerfile
It kinda works in that its fallen into the generic issue of the h2===4.0 error. Will test to see if that can be fixed in the environment.yml or if Docker commands have to be written to uninstall it and install with different version. But thats a problem for tomorrow.